updated getNotebookStructure and sent notebook root location to electron
